number_of_dimensions_t as named_size_t

+++ cc/chart-relax-save-intermediate-layouts.cc
@@ -39,7 +39,7 @@ int main(int argc, char* const argv[])
             if (!args["--no-disconnect-having-few-titers"])
                 disconnected.extend(chart.titers()->having_too_few_numeric_titers());
 
-            auto projection = chart.projections_modify()->new_from_scratch(acmacs::number_of_dimensions_t{args["-d"]}, static_cast<std::string_view>(args["-m"]));
+            auto projection = chart.projections_modify()->new_from_scratch(acmacs::number_of_dimensions_t{static_cast<long>(args["-d"])}, static_cast<std::string_view>(args["-m"]));
             projection->randomize_layout(acmacs::chart::ProjectionModify::randomizer::plain_from_sample_optimization, 2.0);
             acmacs::chart::IntermediateLayouts intermediate_layouts;
             const auto status = projection->relax(acmacs::chart::optimization_options(method, precision, args["--md"]), intermediate_layouts);
